Added a separate print control data structure to ml_struct.h. This widget is unfortunately global.  The widget value is integer. Various routines compare their printing flage value to the widget value to determine if they should print (less than the widget value means print). Other printing flags are double to allow finer print control.

+++ packages/ml/src/Coarsen/ml_aggregate.h
@@ -23,7 +23,7 @@
 typedef struct ML_Aggregate_Struct
 {
    int    ML_id;
-   int    print_flag;
+   double print_flag;
    int    max_coarse_size;             /* maximum size of coarsest grid */
    int    ordering;                    /* natural, random, graph        */
    int    min_nodes_per_aggregate;     /* aggregate size control        */
@@ -121,7 +121,7 @@ extern "C"
 int ML_Aggregate_Create( ML_Aggregate ** );
 int ML_Aggregate_Destroy( ML_Aggregate ** );
 
-int ML_Aggregate_Set_OutputLevel( ML_Aggregate *, int level );
+int ML_Aggregate_Set_OutputLevel( ML_Aggregate *, double level );
 int ML_Aggregate_Set_Reuse(ML_Aggregate *ag);
 
 int ML_Aggregate_Set_MaxLevels( ML_Aggregate *, int level );
